pyproject.toml
# This file is part of "pytest-austin" which is released under GPL.
#
# See file LICENCE or go to http://www.gnu.org/licenses/ for full license
# details.
#
# pytest-austin is a Python wrapper around Austin, the CPython frame stack
# sampler.
#
# Copyright (c) 2018-2020 Gabriele N. Tornetta <phoenix1987@gmail.com>.
# All rights reserved.
#
# This program is free software: you can redistribute it and/or modify
# it under the terms of the GNU General Public License as published by
# the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.
#
# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU General Public License for more details.
# You should have received a copy of the GNU General Public License
# along with this program.  If not, see <http://www.gnu.org/licenses/>.

[tool.poetry]
name = "pytest-austin"
version = "0.1.0"
description = "Austin plugin for pytest"
license = "GPL-3.0-or-later"
authors = ["Gabriele N. Tornetta <phoenix1987@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/P403n1x87/pytest-austin"
repository = "https://github.com/P403n1x87/pytest-austin"
documentation = "https://pytest-austin.readthedocs.io"
keywords = ["performance", "profiling", "testing", "development"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Framework :: Pytest",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: GPLv3",
  "Programming Language :: Python :: 3.6",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
]

[tool.poetry.plugins.pytest11]
cool_plugin = "pytest_austin.plugin"

[tool.poetry.dependencies]
python = ">=3.6"
austin-python = {git = "https://github.com/P403n1x87/austin-python.git", branch = "devel"}
# dataclasses = "*"
# psutil = ">=5.7.0"
# protobuf = "^3.12.2"

[tool.poetry.dev-dependencies]
coverage = {extras = ["toml"], version = "*"}
pytest = ">=5.4.2"
pytest-cov = ">=2.8.1"
nox = "^2020.5.24"
codecov = "^2.1.3"

[tool.poetry.urls]
issues = "https://github.com/P403n1x87/pytest-austin/issues"

[tool.coverage.run]
branch = true
source = ["pytest_austin"]

[tool.coverage.report]
show_missing = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
